// pioneer/packages/joy-media/src/schemas/music/MusicTrack.ts
export type MusicTrackType = {
  classId: number;
  inClassSchemaIndexes: number[];
  id: number;
  title: string;
  artist: string;
  thumbnail: string;
  description?: string;
  language?: LanguageType;
  firstReleased: number;
  genre?: MusicGenreType;
  mood?: MusicMoodType;
  theme?: MusicThemeType;
  links?: string[];
  composerOrSongwriter?: string;
  lyrics?: string;
  object?: MediaObjectType;
  publicationStatus: PublicationStatusType;
  curationStatus?: CurationStatusType;
  explicit: boolean;
  license: ContentLicenseType;
  attribution?: string;
  channelId?: number;
  channel?: ChannelEntity;
};